UPSTREAM: arm64: support __int128 with clang
Commit fb8722735f50 ("arm64: support __int128 on gcc 5+") added support
for arm64 __int128 with gcc with a version-conditional, but neglected to
enable this for clang, which in fact appears to support aarch64 __int128.
This commit therefore enables it if the compiler is clang, using the
same type of makefile conditional used elsewhere in the tree.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit ad40bdafb495f30e5af837e15b3c2a4cb2176e47)
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: utsavbalar1231 <utsavbalar1231@gmail.com>